+++ bld/rc/rc/c/ppalloc.c
@@ -33,9 +33,10 @@
 #include "errors.h"
 #include "layer0.h"
 #include "rcrtns.h"
+#include "preproc.h"
 
-void *PP_Malloc( unsigned size )
-/******************************/
+void *PP_Malloc( size_t size )
+/****************************/
 {
     return( RCALLOC( size ) );
 }
